Major functionality enhancement, including time-based calculation and rendering.
Add advanceFrame() and advanceTimeStep() methods to the FluidSolver class, as well as a QTimer that continually attempts to calculate the next frame for the scene.

Calculations are now based on the estimated maximum velocity within the simulation as well as the user-specified CFL coefficient (currently hardcoded to 2).

Create stubs for the pressure solve, advection, and point movement methods within the FluidSolver class.

Enforce boundary conditions on the fluid velocity field, clamping values at simulation edges to 0.

Enhance documentation throughout the project.  Still no readme file, though!